Improved version

+++ modules/task_3/novikov_d_jarvis/novikov_d_jarvis.h
@@ -7,10 +7,10 @@
 #include <ctime>
 
 struct point {
-    int x;
-    int y;
+    int64_t x;
+    int64_t y;
 
-    point(int x, int y) : x(x), y(y) {}
+    point(int64_t x, int64_t y) : x(x), y(y) {}
 
     point() = default;
